// src/components/neon-crm/types.ts
export type ContactStatus = {
  tenant_id: string;
  label: string;
  color: string;
  icon?: string;
  position: number;
  is_default: boolean;
  created_at: string;
} & Pick<RaRecord, "id">;